linking Run Button to Table and Line Chart
