+++ src/app/line-chart/line-chart.component.html
@@ -1,4 +1,3 @@
 <div class="indiv-summary"><b><span class="bottom-border">$</span></b> in <b><span class="bottom-border">Total US
             food</span></b></div>
-<div class="line-chart" [chart]="lineChart"></div>
-<button class="actual-data" (click)="renderLineChart()">Show Data</button>
+<div class="line-chart" [chart]="lineChartDataService.lineChart"></div>
