+++ src/app/line-chart/line-chart.component.scss
@@ -15,13 +15,3 @@
   margin-top: 16px;
 }
 
-.actual-data {
-  position: absolute;
-  top: 50px;
-  right: 16px;
-  padding: 6px 8px;
-  background-color: blue;
-  border-radius: 4px;
-  color: white;
-  border: none;
-}
